extract geametry objects, light, calc time
